% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_impute.R
\name{base_impute}
\alias{base_impute}
\title{Basic (naive) imputation for missing values in a data.table.}
\usage{
base_impute(dt, type = "median", impute_col, group_by = "")
}
\arguments{
\item{dt}{A data.table with a column of with missing values.}

\item{type}{The type of imputation to be done; options are 'mean', 'median',
and 'zero'.}

\item{impute_col}{The column on which to perform imputation.}

\item{group_by}{The columns to group the imputation method by.}
}
\description{
Basic (naive) imputation for missing values in a data.table.
}
\examples{
base_impute(bee.data, "median", "TOTAL_COUNT", c("MONTH", "DAY"))
}
