% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_arma.R
\name{sparse_arma}
\alias{sparse_arma}
\title{Sparse ARMA Model with Auto-Selected Coefficients}
\usage{
sparse_arma(x, p_max = 10, q_max = 10, pct_ci = 0.95, margin = 0)
}
\arguments{
\item{x}{A time series object}

\item{p_max}{The maximum number of coefficients for the AR part of the model}

\item{q_max}{The maximum number of coefficients for the MA part of the model}

\item{pct_ci}{What percentage of confidence interval to be used for
coefficient consideration.}

\item{margin}{An additional margin to increase sparsity.}
}
\description{
Sparse ARMA Model with Auto-Selected Coefficients
}
\examples{
data <- base_impute(bee.data[1:100], impute_col = "TOTAL_COUNT")
x <- data$IMPUTED_VALS
model <- sparse_arma(x)
}
