% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_ma.R
\name{sparse_ma}
\alias{sparse_ma}
\title{Sparse MA Model with Auto-Selected Coefficients}
\usage{
sparse_ma(x, q_max = 30, pct_ci = 0.95, margin = 0)
}
\arguments{
\item{x}{A time series object}

\item{q_max}{The maximum number of coefficients to be compute}

\item{pct_ci}{What percentage of confidence interval to be used for
coefficient consideration.}

\item{margin}{An additional margin to increase sparsity.}
}
\description{
Sparse MA Model with Auto-Selected Coefficients
}
\examples{
data <- base_impute(bee.data[1:100], impute_col = "TOTAL_COUNT")
x <- data$IMPUTED_VALS
model <- sparse_ma(x)
}
